requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-weaviate>=0.0.3
langgraph>=0.2.0
streamlit>=1.40.0
weaviate-client>=4.9.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.0.0
